\name{suppfam-package}
\alias{suppfam-package}
\alias{suppfam}
\docType{package}
\title{
\packageTitle{suppfam}
}
\description{
\packageDescription{suppfam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{suppfam}
\packageIndices{suppfam}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{suppfam}

Maintainer: \packageMaintainer{suppfam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
